requirements.txt
kubernetes
packaging
prometheus-client
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
